Crate aptos_global_constants

The purpose of this crate is to offer a single source of truth for the definitions of shared constants within the codebase. This is useful because many different components within Aptos often require access to global constant definitions (e.g., Safety Rules, Key Manager, and Secure Storage). To avoid duplicating these definitions across crates (and better allow these constants to be updated in a single location), we define them here.

Constants

APTOS_ROOT_KEY

Definitions of global cryptographic keys (e.g., as held in secure storage)

CONSENSUS_KEY

FULLNODE_NETWORK_KEY

GAS_HEADROOM_DENOMINATOR

GAS_HEADROOM_NUMERATOR

GAS_UNIT_PRICE

GENESIS_WAYPOINT

MAX_GAS_AMOUNT

MIN_PRICE_PER_GAS_UNIT

MOVE_MODULES

OPERATOR_ACCOUNT

OPERATOR_KEY

OWNER_ACCOUNT

OWNER_KEY

SAFETY_DATA

Definitions of global data items (e.g., as held in secure storage)

VALIDATOR_NETWORK_KEY

WAYPOINT

Functions

adjust_gas_headroom

Gas costs are dynamic based on storage, so the simulation values need some headroom applied by the user if using it to estimate gas